+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2024-02-29T12:28:48.032Z\n"
-"PO-Revision-Date: 2024-02-29T12:28:48.032Z\n"
+"POT-Creation-Date: 2024-02-29T12:43:22.968Z\n"
+"PO-Revision-Date: 2024-02-29T12:43:22.968Z\n"
 
 msgid "Choose one or more dates..."
 msgstr "Choose one or more dates..."
@@ -314,6 +314,24 @@ msgstr "Enroll in a program by selecting a program from the top bar."
 msgid "An error has occurred. See log for details"
 msgstr "An error has occurred. See log for details"
 
+msgid "{{programStageName}} completed"
+msgstr "{{programStageName}} completed"
+
+msgid "Would you like to complete the enrollment and all active events as well?"
+msgstr "Would you like to complete the enrollment and all active events as well?"
+
+msgid "Yes, complete enrollment and events"
+msgstr "Yes, complete enrollment and events"
+
+msgid "Complete enrollment only"
+msgstr "Complete enrollment only"
+
+msgid "Would you like to complete the enrollment?"
+msgstr "Would you like to complete the enrollment?"
+
+msgid "Complete enrollment"
+msgstr "Complete enrollment"
+
 msgid "A duplicate exists (but there were some errors, see log for details"
 msgstr "A duplicate exists (but there were some errors, see log for details"
 
@@ -354,24 +372,6 @@ msgstr "Indicators"
 
 msgid "Warnings"
 msgstr "Warnings"
-
-msgid "{{programStageName}} completed"
-msgstr "{{programStageName}} completed"
-
-msgid "Would you like to complete the enrollment and all active events as well?"
-msgstr "Would you like to complete the enrollment and all active events as well?"
-
-msgid "Yes, complete enrollment and events"
-msgstr "Yes, complete enrollment and events"
-
-msgid "Complete enrollment only"
-msgstr "Complete enrollment only"
-
-msgid "Would you like to complete the enrollment?"
-msgstr "Would you like to complete the enrollment?"
-
-msgid "Complete enrollment"
-msgstr "Complete enrollment"
 
 msgid "Generate new event"
 msgstr "Generate new event"
