+++ src/core_modules/capture-core/components/DataEntries/common/trackerEvent/index.js
