+++ src/core_modules/capture-core/components/DataEntries/common/trackerEvent/withAskToCompleteEnrollment/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { withAskToCompleteEnrollment } from './withAskToCompleteEnrollment';
