+++ src/core_modules/capture-core/components/DataEntry/index.js
@@ -15,7 +15,6 @@ export { withErrorOutput } from './dataEntryOutput/withErrorOutput';
 export { withIndicatorOutput } from './dataEntryOutput/withIndicatorOutput';
 export { withCleanUp } from './withCleanUp';
 export { withAskToCreateNew } from './withAskToCreateNew';
-export { withAskToCompleteEnrollment } from './withAskToCompleteEnrollment';
 
 // misc
 export { inMemoryFileStore } from './file/inMemoryFileStore';
